+++ operator/config/crd/bases/forklift.konveyor.io_plans.yaml
@@ -733,6 +733,9 @@ spec:
                           description: The VM Namespace Only relevant for an openshift
                             source.
                           type: string
+                        ovfConfig:
+                          description: virt-v2v ovf file
+                          type: string
                         phase:
                           description: Phase
                           type: string
